rename: searchShops.ts를 app/user/shop/component 로 이동
